requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
seaborn
matplotlib
PySastrawi
